LayoutTests/fast/mediastream/MediaStreamTrack.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>
            var mediaStream;
            var videoTrack;
            var videoTrack2;
            var audioTrack;
            
            function error()
            {
                testFailed('Stream generation failed.');
                finishJSTest();
            }

            function getUserMedia(constraints, callback)
            {
                try {
                    navigator.webkitGetUserMedia(constraints, callback, error);
                } catch (e) {
                    testFailed('webkitGetUserMedia threw exception :' + e);
                    finishJSTest();
                }
            }
            
            function onTrackEnded()
            {
                testPassed('Track onended callback succeeded.');
                shouldBeEqualToString('videoTrack.readyState', 'ended');
                debug("");
                finishJSTest();
            }

            function onTrackStarted()
            {
                testPassed('Track onstarted callback succeeded.');
                shouldBeEqualToString('videoTrack.readyState', 'live');
            }

            function onTrackOverConstrained()
            {
                testPassed('Track onoverconstrained callback succeeded.');
                finishJSTest();
            }

            function onTrackUnmute()
            {
                testPassed('Track onunmute callback succeeded.');
                shouldBeEqualToString('videoTrack.readyState', 'live');
                videoTrack.stop();
            }

            function onTrackMute()
            {
                testPassed('Track onmute callback succeeded.');
                shouldBeEqual('videoTrack.muted', true);
            }

            function logCapabilityRange(name, range)
            {
                debug("  " + name);
                debug("   min : " + range.min);
                debug("   max : " + range.max);
                debug("   supported : " + range.supported);
            }

            function logCapabilityList(name, list)
            {
                debug("  " + name);
                for (i = 0; i < list.length; i++)
                debug("    " + i + " : " + list[i]);
            }

            function checkCapabilities(track)
            {
                debug("checking track.capabilities()");
                capabilities = track.capabilities();
                logCapabilityList("sourceId", capabilities.sourceId);
                if (track.kind == "video") {
                    logCapabilityList("sourceType", capabilities.sourceType);
                    logCapabilityRange("width", capabilities.width);
                    logCapabilityRange("height", capabilities.height);
                    logCapabilityRange("frameRate", capabilities.frameRate);
                    logCapabilityRange("aspectRatio", capabilities.aspectRatio);
                    logCapabilityList("facingMode", capabilities.facingMode);
                } else
                    logCapabilityRange("volume", capabilities.volume);
            }

            function checkStates(track)
            {
                debug("checking track.states()");
                capabilities = track.capabilities();
                states = track.states();
                shouldBe('states.sourceId', 'capabilities.sourceId[0]');
                if (track.kind == "video") {
                    shouldBeEqualToString('states.sourceType', 'camera');
                    debug("  width : " + states.width);
                    debug("  height : " + states.height);
                    debug("  frameRate : " + states.frameRate);
                    debug("  aspectRatio : " + states.aspectRatio);
                    debug("  facingMode : " + states.facingMode);
                } else {
                    shouldBeEqualToString('states.sourceType', 'microphone');
                    debug("  volume : " + states.volume);
                }
            }
            
            function checkTrack(t)
            {
                track = t;
                shouldBeEqualToString('track.readyState', 'live');
                checkCapabilities(track);
                checkStates(track);
            }

            function gotStream(stream)
            {
                mediaStream = stream;
                testPassed('getUserMedia succeeded.');
            
                debug("");
                evalAndLog("videoTrack = mediaStream.getVideoTracks()[0]");
                checkTrack(videoTrack);

                debug("");
                evalAndLog("audioTrack = mediaStream.getAudioTracks()[0]");
                checkTrack(audioTrack);

                videoTrack.onmute = onTrackMute;
                videoTrack.onunmute = onTrackUnmute;
                videoTrack.onended = onTrackEnded;
                videoTrack.onstarted = onTrackStarted;
                videoTrack.onoverconstrained = onTrackOverConstrained;
            
                debug("");
                evalAndLog("videoTrack2 = videoTrack.clone()");
                checkTrack(videoTrack2);
                shouldNotBe('videoTrack.id', 'videoTrack2.id');
                shouldBe('videoTrack.kind', 'videoTrack2.kind');
                shouldBe('videoTrack.label', 'videoTrack2.label');
                shouldBe('videoTrack.muted', 'videoTrack2.muted');
                shouldBe('videoTrack.enabled', 'videoTrack2.enabled');
                shouldBe('videoTrack.readyState', 'videoTrack2.readyState');

                debug("");
                evalAndLog("videoTrack.stop()");
            }

            function start()
            {
                description("Tests MediaStreamTrack callbacks.");
                
                // Note that the below behaviour doesn't reflect how it works outside of LayoutTests.
                // The underlying mock is modified to trigger the events when certain functions are called.
                // This modified behaviour allows us to test the MediaStreamTrack class properly.
                getUserMedia({audio:true, video:true}, gotStream);
            }

            window.jsTestIsAsync = true;
            window.successfullyParsed = true;
        </script>
    </head>
    <body onload="start()">
        <p id="description"></p>
        <div id="console"></div>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
